+++ README.md
@@ -1,4 +1,4 @@
-bit.ly URL shortening plugin
+# bit.ly URL shortening plugin
 
 You must set up the API keys before you can use this. This is done on the settings
 page for the bit.ly plugin.
