+++ manifest.xml
@@ -1,15 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <plugin_manifest xmlns="http://www.elgg.org/plugin_manifest/1.8">
-	<name>bit.ly URL Shortener</name>
-	<author>Cash Costello</author>
-	<version>1.0</version>
-	<category>utility</category>
-	<description>A URL shortener using the bit.ly service</description>
-	<website>http://cashcostello.com/</website>
-	<copyright>Cash Costello 2011</copyright>
-	<license>GNU General Public License version 2</license>
-	<requires>
-		<type>elgg_release</type>
-		<version>1.8</version>
-	</requires>
+    <name>bit.ly URL Shortener</name>
+    <id>bitly</id>
+    <author>Cash Costello</author>
+    <author>Matt Beckett</author>
+    <version>1.1</version>
+    <category>utility</category>
+    <description>A URL shortener using the bit.ly service</description>
+    <website>http://cashcostello.com/</website>
+    <copyright>Cash Costello 2011</copyright>
+    <license>GNU General Public License version 2</license>
+    <requires>
+        <type>elgg_release</type>
+        <version>1.9</version>
+    </requires>
 </plugin_manifest>
