+++ views/default/css/bitly.php
@@ -1,8 +1,3 @@
-<?php
-/**
- * bit.ly extensions to Elgg's CSS
- */
-?>
 
 .bitly-wrapper {
 	text-align: right;
